requirements.txt
google-generativeai
python-dotenv
asyncpg
pytest
asyncio
